Moved `PredicateKotlinName` and `HavocKotlinName` out of `KotlinName` into `SymbolicName`
